euclid-data-platform/edap/requirements/base.txt
Django==3.2
pandas==1.3.0
numpy==1.21.0
elasticsearch==7.13.1
mysqlclient==2.0.3
django_elasticsearch_dsl==7.2.0
djangorestframework==3.12.4
sqlalchemy==1.4.20
click==8.0.1
python-igraph==0.9.6
scipy==1.7.0
tqdm==4.61.2
fastparquet==0.6.3
psutil==5.8.0
paramiko==2.7.2
sklearn
drf-yasg>=1
django-cors-headers==3.6.0
gunicorn==20.1.0
cryptography==3.4.7
gensim==4.1.0
python-Levenshtein==0.12.2
networkx==2.6.3